web-spacedj/checker/requirements.txt
annotated-types==0.7.0
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
checklib==0.7.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dtw-python==1.5.3
fonttools==4.54.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
matplotlib==3.9.2
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
pycparser==2.22
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
